Simplify the Loadbalancing section
Previously, the loadbalancing section required multi-threaded and
multi-server capability. At the IETF 109 session, we concluded that
the multi-threaded aspects were at most an implementation detail and
likely not required. However, there was support for the
multi-session/multi-server capability, although likely as an
extension.

This commit updates the text to mention the mutli-session capability
as desirable, but explicitly calls out that implementing it may be an
extension. This addresses
ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#9 and
ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#11 by making this
optional.