Clarify server requesting an IP address from the client
This change attempts to address
ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#3, by specifying the
server may request an IP in the client's range. This is not required
for operation of network-to-network proxying, unlike the document
previously claimed.

Logically, in network-to-network connectivity, the only difference
between the client and the server is which one initiates the
connection. There is no requirement that the server is making
resources available to the client, but rather, the client could be
making things available to the server. In such an operating mode the
server would be assigned an address in the client's range.

One potential use case could be an enterprise service behind a NAT or
firewall being made available to a network-level cloud load balancer.