Clarify packet formats to permit extensions
Previously, the text wrote the data transports must transmit packets
in their unmodified entirety, precluding extensions. This commit
updates the text to note extensions of the data formats are
permissible.

This addresses ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#10 but
does not change the terminology as there does not seem to be consensus
yet. This also addresses
ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#13 by removing the
only-unmodified requirement.