Replace the load balancing section with a Statefulness requirement
During discussion with @DavidSchinazi, we realized that the load
balancing requirement was actually standing-in for a more relevant
requirement that had been thus far implicit. The true requirement is
around statefulness: the state in both the client and server should be
kept minimal, as this makes it easier to operate the protocol, as well
as re-connect in the event of failure.

We believe the load balancing capability can be implemented with an
extension and needs no further requirements beyond the extensibility
details already included. The load balancing section has been removed.

This should address ietf-wg-masque/draft-ietf-masque-ip-proxy-reqs#9.